Move decompose classes to TvManiacKit package.
